requirements.txt
async-generator==1.10
attrs==22.1.0
behave==1.2.6
certifi==2022.12.7
cffi==1.15.1
h11==0.14.0
idna==3.4
outcome==1.2.0
parse-type==0.6.0
parse==1.19.0
pip==21.3.1
pycparser==2.21
PySocks==1.7.1
selenium==4.7.2
setuptools==60.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.9.2
trio==0.22.0
urllib3==1.26.13
wheel==0.37.1
wsproto==1.2.0
python-dotenv~=0.21.0
google-cloud-storage==2.7.0